not final

+++ src/main/java/dao/DAO.java
@@ -1,14 +1,12 @@
 package dao;
 
-import entity.Shops;
-
 import java.util.List;
 
 public interface DAO <T> {
     T add(T t);
     boolean delete(T t);
     T finByName(String k);
-    T finByID(int id);
+    T finByID(long id);
     List<T> findAll();
     boolean update(T t);
